updater: optimization and re-structuring
-better organization of the files
-saves more memory by organizing like this
